opt vector, add biginteger add/sub/mul

+++ include/wjr/biginteger/biginteger.hpp
@@ -9,47 +9,49 @@ namespace wjr {
 
 namespace biginteger_details {
 
-WJR_CONST inline uint32_t abs_size(uint32_t n) { return n & 0x7FFFFFFF; }
-WJR_CONST inline uint32_t mask_sign(uint32_t n) { return n & 0x80000000; }
+WJR_CONST WJR_CONSTEXPR inline uint32_t abs_size(uint32_t n) { return n & 0x7FFFFFFF; }
+WJR_CONST WJR_CONSTEXPR inline uint32_t negative_mask() { return 0x80000000; }
+WJR_CONST WJR_CONSTEXPR inline uint32_t mask_sign(uint32_t n) {
+    return n & negative_mask();
+}
 
 } // namespace biginteger_details
 
 class default_biginteger_size_reference : noncopyable {
 public:
     default_biginteger_size_reference() = delete;
-    WJR_CONSTEXPR20 explicit default_biginteger_size_reference(uint32_t &size) noexcept
-        : m_size(size) {}
+    explicit default_biginteger_size_reference(uint32_t &size) noexcept : m_size(size) {}
     ~default_biginteger_size_reference() = default;
+    default_biginteger_size_reference(default_biginteger_size_reference&&) = default;
+    default_biginteger_size_reference& operator=(default_biginteger_size_reference&&) = default;
 
-    WJR_CONSTEXPR20 default_biginteger_size_reference &operator=(uint32_t size) noexcept {
+    default_biginteger_size_reference &operator=(uint32_t size) noexcept {
         using namespace biginteger_details;
         WJR_ASSUME(!mask_sign(size));
         m_size = mask_sign(m_size) | size;
         return *this;
     }
 
-    WJR_PURE WJR_CONSTEXPR20 operator uint32_t() const noexcept {
+    WJR_PURE operator uint32_t() const noexcept {
         return biginteger_details::abs_size(m_size);
     }
 
-    WJR_CONSTEXPR20 default_biginteger_size_reference &operator++() noexcept {
+    default_biginteger_size_reference &operator++() noexcept {
         ++m_size;
         return *this;
     }
 
-    WJR_CONSTEXPR20 default_biginteger_size_reference &operator--() noexcept {
+    default_biginteger_size_reference &operator--() noexcept {
         --m_size;
         return *this;
     }
 
-    WJR_CONSTEXPR20 default_biginteger_size_reference &
-    operator+=(uint32_t size) noexcept {
+    default_biginteger_size_reference &operator+=(uint32_t size) noexcept {
         m_size += size;
         return *this;
     }
 
-    WJR_CONSTEXPR20 default_biginteger_size_reference &
-    operator-=(uint32_t size) noexcept {
+    default_biginteger_size_reference &operator-=(uint32_t size) noexcept {
         m_size -= size;
         return *this;
     }
@@ -90,34 +92,47 @@ public:
     default_biginteger_vector_storage() noexcept = default;
 
     template <typename _Alloc>
-    WJR_CONSTEXPR20 default_biginteger_vector_storage(_Alloc &&al) noexcept
+    default_biginteger_vector_storage(_Alloc &&al) noexcept
         : m_pair(std::forward<_Alloc>(al), {}) {}
 
     template <typename _Alloc>
-    WJR_CONSTEXPR20 default_biginteger_vector_storage(_Alloc &&al, size_type size,
-                                                      size_type capacity,
-                                                      in_place_reallocate_t) noexcept
+    default_biginteger_vector_storage(_Alloc &&al, size_type size, size_type capacity,
+                                      in_place_reallocate_t) noexcept
         : m_pair(std::forward<_Alloc>(al), {}) {
         uninitialized_construct(size, capacity);
     }
 
     ~default_biginteger_vector_storage() noexcept = default;
 
-    WJR_PURE WJR_CONSTEXPR20 _Alty &get_allocator() noexcept { return m_pair.first(); }
-    WJR_PURE WJR_CONSTEXPR20 const _Alty &get_allocator() const noexcept {
-        return m_pair.first();
-    }
+    WJR_PURE _Alty &get_allocator() noexcept { return m_pair.first(); }
+    WJR_PURE const _Alty &get_allocator() const noexcept { return m_pair.first(); }
 
-    WJR_CONSTEXPR20 void destroy() noexcept {
+    void destroy() noexcept {
+        if (WJR_BUILTIN_CONSTANT_P(data() == nullptr) && data() == nullptr) {
+            return;
+        }
+
         const size_type __size = size();
+
+        if (WJR_BUILTIN_CONSTANT_P(__size == 0) && __size == 0) {
+            return;
+        }
+
         destroy_n_using_allocator(data(), __size, get_allocator());
     }
 
-    WJR_CONSTEXPR20 void deallocate() noexcept {
-        get_allocator().deallocate(data(), capacity());
+    void destroy_and_deallocate() noexcept {
+        if (WJR_BUILTIN_CONSTANT_P(capacity() == 0) && capacity() == 0) {
+            return;
+        }
+
+        if (data()) {
+            destroy();
+            get_allocator().deallocate(data(), capacity());
+        }
     }
 
-    WJR_CONSTEXPR20 void uninitialized_construct(size_type _size, size_type capacity) {
+    void uninitialized_construct(size_type _size, size_type capacity) {
         auto &al = get_allocator();
         auto &m_storage = __get_data();
         m_storage.m_data = al.allocate(capacity);
@@ -125,8 +140,7 @@ public:
         m_storage.m_capacity = capacity;
     }
 
-    WJR_CONSTEXPR20 void
-    take_storage(default_biginteger_vector_storage &&other) noexcept {
+    void take_storage(default_biginteger_vector_storage &&other) noexcept {
         auto &other_storage = other.__get_data();
         auto &__storage = __get_data();
         __storage.m_data = other_storage.m_data;
@@ -136,37 +150,33 @@ public:
         other_storage = {};
     }
 
-    WJR_CONSTEXPR20 void swap_storage(default_biginteger_vector_storage &other) noexcept {
+    void swap_storage(default_biginteger_vector_storage &other) noexcept {
         std::swap(__get_data(), other.__get_data());
     }
 
-    WJR_PURE WJR_CONSTEXPR20 default_biginteger_size_reference size() noexcept {
+    WJR_PURE default_biginteger_size_reference size() noexcept {
         return default_biginteger_size_reference(__get_data().m_size);
     }
-    WJR_PURE WJR_CONSTEXPR20 size_type size() const noexcept {
+    WJR_PURE size_type size() const noexcept {
         return biginteger_details::abs_size(__get_data().m_size);
     }
-    WJR_PURE WJR_CONSTEXPR20 size_type capacity() const noexcept {
-        return __get_data().m_capacity;
-    }
+    WJR_PURE size_type capacity() const noexcept { return __get_data().m_capacity; }
 
-    WJR_PURE WJR_CONSTEXPR20 pointer data() noexcept { return __get_data().m_data; }
-    WJR_PURE WJR_CONSTEXPR20 const_pointer data() const noexcept {
-        return __get_data().m_data;
-    }
+    WJR_PURE pointer data() noexcept { return __get_data().m_data; }
+    WJR_PURE const_pointer data() const noexcept { return __get_data().m_data; }
 
     // extension
 
-    WJR_PURE WJR_CONSTEXPR20 int32_t get_ssize() const noexcept {
-        return __get_data().m_size;
+    WJR_PURE int32_t get_ssize() const noexcept { return __get_data().m_size; }
+    void set_ssize(int32_t size) noexcept {
+        WJR_ASSERT(biginteger_details::abs_size(size) <= capacity());
+        __get_data().m_size = size;
     }
-    WJR_CONSTEXPR20 void set_ssize(int32_t size) noexcept { __get_data().m_size = size; }
+    void set_nullptr() noexcept { __get_data().m_data = nullptr; }
 
 private:
-    WJR_PURE WJR_CONSTEXPR20 data_type &__get_data() noexcept { return m_pair.second(); }
-    WJR_PURE WJR_CONSTEXPR20 const data_type &__get_data() const noexcept {
-        return m_pair.second();
-    }
+    WJR_PURE data_type &__get_data() noexcept { return m_pair.second(); }
+    WJR_PURE const data_type &__get_data() const noexcept { return m_pair.second(); }
 
     compressed_pair<_Alty, data_type> m_pair;
 };
@@ -182,9 +192,9 @@ inline uint32_t normalize(uint64_t *ptr, uint32_t n) {
     return reverse_find_not_n(ptr, 0, n);
 }
 
-template <bool xsign>
-WJR_CONST WJR_CONSTEXPR20 inline int32_t conditional_negate(int32_t x) noexcept {
-    return xsign ? (x & 1ull << 31) : x;
+WJR_CONST WJR_CONSTEXPR inline int32_t conditional_negate(bool xsign,
+                                                          int32_t x) noexcept {
+    return xsign ? (x | biginteger_details::negative_mask()) : x;
 }
 
 } // namespace biginteger_details
@@ -217,32 +227,133 @@ public:
     basic_biginteger &operator=(basic_biginteger &&other) = default;
     ~basic_biginteger() = default;
 
-    WJR_PURE WJR_CONSTEXPR20 pointer data() noexcept { return m_vec.data(); }
-    WJR_PURE WJR_CONSTEXPR20 const_pointer data() const noexcept { return m_vec.data(); }
+    explicit basic_biginteger(const allocator_type &al) : m_vec(al) {}
 
-    WJR_PURE WJR_CONSTEXPR20 size_type size() const noexcept { return m_vec.size(); }
-    WJR_PURE WJR_CONSTEXPR20 size_type capacity() const noexcept {
-        return m_vec.capacity();
+    template <typename Iter, std::enable_if_t<is_iterator_v<Iter>, int> = 0>
+    basic_biginteger(Iter first, Iter last, const allocator_type &al)
+        : m_vec(first, last, al) {
+        __check_high_bit();
     }
+
+    basic_biginteger(const basic_biginteger &other, const allocator_type &al)
+        : m_vec(other.m_vec, al) {}
+
+    basic_biginteger(basic_biginteger &&other, const allocator_type &al)
+        : m_vec(std::move(other.m_vec), al) {}
+
+    basic_biginteger(std::initializer_list<value_type> il, const allocator_type &al)
+        : m_vec(il, al) {
+        __check_high_bit();
+    }
+
+    basic_biginteger(size_type n, in_place_default_construct_t, const allocator_type &al)
+        : m_vec(n, in_place_default_construct, al) {}
+
+    basic_biginteger &operator=(std::initializer_list<value_type> il) {
+        m_vec = il;
+        __check_high_bit();
+        return *this;
+    }
+
+    template <typename Iter, std::enable_if_t<is_iterator_v<Iter>, int> = 0>
+    void assign(Iter first, Iter last) {
+        m_vec.assign(first, last);
+        __check_high_bit();
+    }
+
+    void assign(std::initializer_list<value_type> il) {
+        m_vec.assign(il);
+        __check_high_bit();
+    }
+
+    allocator_type &get_allocator() noexcept { return m_vec.get_allocator(); }
+    const allocator_type &get_allocator() const noexcept { return m_vec.get_allocator(); }
+
+    reference at(size_type pos) { return m_vec.at(pos); }
+    const_reference at(size_type pos) const { return m_vec.at(pos); }
+
+    reference operator[](size_type pos) { return m_vec[pos]; }
+    const_reference operator[](size_type pos) const { return m_vec[pos]; }
+
+    reference front() { return m_vec.front(); }
+    const_reference front() const { return m_vec.front(); }
+
+    reference back() { return m_vec.back(); }
+    const_reference back() const { return m_vec.back(); }
+
+    WJR_PURE pointer data() noexcept { return m_vec.data(); }
+    WJR_PURE const_pointer data() const noexcept { return m_vec.data(); }
+
+    WJR_PURE iterator begin() noexcept { return m_vec.begin(); }
+    WJR_PURE const_iterator begin() const noexcept { return m_vec.begin(); }
+
+    WJR_PURE iterator end() noexcept { return m_vec.end(); }
+    WJR_PURE const_iterator end() const noexcept { return m_vec.end(); }
+
+    WJR_PURE const_iterator cbegin() const noexcept { return m_vec.cbegin(); }
+
+    WJR_PURE const_iterator cend() const noexcept { return m_vec.cend(); }
+
+    WJR_PURE reverse_iterator rbegin() noexcept { return m_vec.rbegin(); }
+    WJR_PURE const_reverse_iterator rbegin() const noexcept { return m_vec.rbegin(); }
+
+    WJR_PURE reverse_iterator rend() noexcept { return m_vec.rend(); }
+    WJR_PURE const_reverse_iterator rend() const noexcept { return m_vec.rend(); }
+
+    WJR_PURE const_reverse_iterator crbegin() const noexcept { return m_vec.crbegin(); }
+    WJR_PURE const_reverse_iterator crend() const noexcept { return m_vec.crend(); }
+
+    WJR_PURE bool empty() const noexcept { return m_vec.empty(); }
+    WJR_PURE size_type size() const noexcept { return m_vec.size(); }
+    WJR_PURE size_type capacity() const noexcept { return m_vec.capacity(); }
+
+    WJR_CONST static size_type get_growth_capacity(size_type old_capacity,
+                                                   size_type new_size) noexcept {
+        return vector_type::get_growth_capacity(old_capacity, new_size);
+    }
+
+    void reserve(size_type new_capacity) { m_vec.reserve(new_capacity); }
+
+    void shrink_to_fit() { m_vec.shrink_to_fit(); }
+
+    void clear() { m_vec.clear(); }
+
+    void swap(basic_biginteger &other) noexcept { m_vec.swap(other.m_vec); }
 
     /**
      * @brief Get the sign of biginteger
-
      *
      * @return
      * false : if the biginteger is positive \n
      * true : if the biginteger is negative
      */
-    WJR_PURE WJR_CONSTEXPR20 bool get_sign() const noexcept { return __get_ssize() < 0; }
+    WJR_PURE bool get_sign() const noexcept { return get_ssize() < 0; }
 
-    WJR_CONSTEXPR20 void reserve(size_type new_capacity) { m_vec.reserve(new_capacity); }
+    WJR_PURE int32_t get_ssize() const { return __get_storage().get_ssize(); }
+    void set_ssize(int32_t new_size) { __get_storage().set_ssize(new_size); }
 
-    WJR_CONSTEXPR20 void resize(size_type new_size) { m_vec.resize(new_size); }
-    WJR_CONSTEXPR20 void resize(size_type new_size, value_type value) {
-        m_vec.resize(new_size, value);
+    friend bool operator==(const basic_biginteger &lhs, const basic_biginteger &rhs) {
+        return lhs.m_vec == rhs.m_vec;
     }
-    WJR_CONSTEXPR20 void resize(size_type new_size, in_place_default_construct_t) {
-        m_vec.resize(new_size, in_place_default_construct);
+
+    friend bool operator!=(const basic_biginteger &lhs, const basic_biginteger &rhs) {
+        return lhs.m_vec != rhs.m_vec;
+    }
+
+    friend bool operator<(const basic_biginteger &lhs, const basic_biginteger &rhs) {
+        return lhs.m_vec < rhs.m_vec;
+    }
+
+    friend bool operator>(const basic_biginteger &lhs, const basic_biginteger &rhs) {
+        return lhs.m_vec > rhs.m_vec;
+    }
+
+    friend bool operator<=(const basic_biginteger &lhs, const basic_biginteger &rhs) {
+        return lhs.m_vec <= rhs.m_vec;
+    }
+
+    friend bool operator>=(const basic_biginteger &lhs, const basic_biginteger &rhs) {
+        return lhs.m_vec >= rhs.m_vec;
     }
 
     friend void add(basic_biginteger &dst, const basic_biginteger &lhs,
@@ -255,24 +366,29 @@ public:
         __addsub<true>(&dst, &lhs, &rhs);
     }
 
-private:
-    WJR_PURE WJR_CONSTEXPR20 storage_type &__get_storage() noexcept {
-        return m_vec.get_storage();
+    friend void mul(basic_biginteger &dst, const basic_biginteger &lhs,
+                    const basic_biginteger &rhs) {
+        __mul(&dst, &lhs, &rhs);
     }
-    WJR_PURE WJR_CONSTEXPR20 const storage_type &__get_storage() const noexcept {
+
+private:
+    WJR_PURE storage_type &__get_storage() noexcept { return m_vec.get_storage(); }
+    WJR_PURE const storage_type &__get_storage() const noexcept {
         return m_vec.get_storage();
     }
 
-    WJR_PURE WJR_CONSTEXPR20 int32_t __get_ssize() const noexcept {
-        return __get_storage().get_ssize();
-    }
-    WJR_CONSTEXPR20 void __set_ssize(int32_t size) noexcept {
-        __get_storage().set_ssize(size);
+    void __set_nullptr() noexcept { __get_storage().set_nullptr(); }
+
+    void __check_high_bit() const {
+        WJR_ASSERT(size() == 0 || back() != 0, "biginteger should not have leading zero");
     }
 
     template <bool xsign>
     static void __addsub(basic_biginteger *dst, const basic_biginteger *lhs,
                          const basic_biginteger *rhs);
+
+    static void __mul(basic_biginteger *dst, const basic_biginteger *lhs,
+                      const basic_biginteger *rhs);
 
     vector_type m_vec;
 };
@@ -281,14 +397,19 @@ using biginteger =
     basic_biginteger<default_biginteger_vector_storage<std::allocator<uint64_t>>>;
 
 template <typename Storage>
+void swap(basic_biginteger<Storage> &lhs, basic_biginteger<Storage> &rhs) noexcept {
+    lhs.swap(rhs);
+}
+
+template <typename Storage>
 template <bool xsign>
 void basic_biginteger<Storage>::__addsub(basic_biginteger *dst,
                                          const basic_biginteger *lhs,
                                          const basic_biginteger *rhs) {
     using namespace biginteger_details;
 
-    int32_t lssize = lhs->__get_ssize();
-    int32_t rssize = conditional_negate<xsign>(rhs->__get_ssize());
+    int32_t lssize = lhs->get_ssize();
+    int32_t rssize = conditional_negate(xsign, rhs->get_ssize());
     uint32_t lusize = abs_size(lssize);
     uint32_t rusize = abs_size(rssize);
 
@@ -300,10 +421,10 @@ void basic_biginteger<Storage>::__addsub(basic_biginteger *dst,
 
     dst->reserve(lusize + 1);
 
-    int32_t dssize;
     auto dp = dst->data();
     auto lp = lhs->data();
     auto rp = rhs->data();
+    int32_t dssize;
 
     // different sign
     if ((lssize ^ rssize) < 0) {
@@ -327,7 +448,89 @@ void basic_biginteger<Storage>::__addsub(basic_biginteger *dst,
         }
     }
 
-    dst->__set_ssize(dssize);
+    dst->set_ssize(dssize);
+}
+
+template <typename Storage>
+void basic_biginteger<Storage>::__mul(basic_biginteger *dst, const basic_biginteger *lhs,
+                                      const basic_biginteger *rhs) {
+    using namespace biginteger_details;
+
+    uint32_t lusize = lhs->get_ssize();
+    uint32_t rusize = rhs->get_ssize();
+    uint32_t mask = mask_sign(lusize ^ rusize);
+    lusize = abs_size(lusize);
+    rusize = abs_size(rusize);
+
+    if (lusize < rusize) {
+        std::swap(lhs, rhs);
+        std::swap(lusize, rusize);
+    }
+
+    if (WJR_UNLIKELY(rusize == 0)) {
+        dst->set_ssize(0);
+        return;
+    }
+
+    int32_t dssize;
+
+    if (rusize == 1) {
+        dst->reserve(lusize + 1);
+        auto cf = mul_1(dst->data(), lhs->data(), lusize, rhs->front());
+        dssize = mask | (lusize + (cf != 0));
+        if (cf != 0) {
+            (*dst)[lusize] = cf;
+        }
+        dst->set_ssize(dssize);
+        return;
+    }
+
+    dssize = lusize + rusize;
+
+    auto dp = dst->data();
+    auto lp = (pointer)lhs->data();
+    auto rp = (pointer)rhs->data();
+
+    union tmp_t {
+        tmp_t() { biginteger.__set_nullptr(); }
+        ~tmp_t() {}
+        std::aligned_storage_t<sizeof(uint64_t), alignof(uint64_t)> storage;
+        basic_biginteger biginteger;
+    } temp;
+
+    unique_stack_allocator stkal(math_details::stack_alloc);
+
+    if (dst->capacity() < dssize) {
+        new (&temp.biginteger)
+            basic_biginteger(dssize, in_place_default_construct, dst->get_allocator());
+        dp = temp.biginteger.data();
+    } else {
+        if (dp == lp) {
+            lp = (pointer)stkal.allocate(lusize);
+            if (dp == rp) {
+                rp = lp;
+            }
+            std::copy_n(dp, lusize, lp);
+        } else if (dp == rp) {
+            rp = (pointer)stkal.allocate(rusize);
+            std::copy_n(dp, rusize, rp);
+        }
+    }
+
+    if (WJR_UNLIKELY(lp == rp)) {
+        sqr(dp, lp, lusize);
+    } else {
+        mul_s(dp, lp, lusize, rp, rusize);
+    }
+
+    auto cf = dp[dssize - 1] == 0;
+    dssize = mask | (dssize - cf);
+
+    if (temp.biginteger.data() != nullptr) {
+        *dst = std::move(temp.biginteger);
+    }
+
+    dst->set_ssize(dssize);
 }
 
 } // namespace wjr
